🏗️ Architect: Modularized altitude plotting logic
Refactored the monolithic `apts/plotting/altitude.py` into a modular package `apts/plotting/altitude/`.

Key changes:
- Created `apts/plotting/altitude/` directory.
- Extracted Messier plotting logic into `messier.py` and planetary plotting logic into `planets.py`.
- Decomposed complex functions in both modules to reduce cyclomatic complexity (now all below 10).
- Maintained the public API via re-exports in `__init__.py`.
- Updated unit test mocks in `tests/unit/test_observations.py` to reflect the new package structure.
- Compiled translations to ensure localized plotting tests pass.

This improvement enhances maintainability and readability by following the Single Responsibility Principle.
