+++ apts/plotting/altitude/__init__.py
@@ -1,0 +1,4 @@
+from .messier import generate_plot_messier
+from .planets import generate_plot_planets
+
+__all__ = ["generate_plot_messier", "generate_plot_planets"]
